Bpod ephys extraction consolidation
